// apps/desktop/src/shared/api/generated/ShareableSettingsDto.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

/**
 * A mod's settings files that could be included in a bundle.
 */
export type ShareableSettingsDto = {
  unique_id: string;
  name: string;
  files: Array<string>;
  /**
   * What the files might reveal if shared.
   */
  warnings: Array<string>;
};